pyproject.toml
[tool.poetry]
name = "pyttnet"
version = "0.0.1"
description = "Updated implementation of https://arxiv.org/pdf/2004.09927.pdf"
authors = ["onlinejudge95 <onlinejudge95@gmail.com>"]
readme = "README.md"
packages = [{include = "src/ttnet"}]

[tool.poetry.scripts]
ttnet = "src.ttnet.entrypoint:cli"

[tool.poetry.dependencies]
python = "3.12.4"
click = "8.1.7"
click-option-group = "0.5.6"
torch = { version = "2.4.0", source = "pytorch-cpu" }
numpy = "1.26.4"
tensorboard = "2.17.0"
opencv-python-headless = "4.9.0.80"
scikit-learn = "1.5.1"
pyturbojpeg = "1.7.5"
tqdm = "4.66.5"

[tool.poetry.group.dev.dependencies]
ruff = "0.5.5"
mypy = {extras = ["reports"], version = "1.11.1"}
pytest = "8.3.2"
pytest-cov = "5.0.0"
pytest-mock = "3.14.0"

[[tool.poetry.source]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
include = ["*.py"]
line-length = 120
output-format = "grouped"
respect-gitignore = true
src = ["src"]

[tool.ruff.format]
docstring-code-format = true
line-ending = "native"
quote-style = "double"
skip-magic-trailing-comma = true

[tool.mypy]
files = "**/*.py"

disallow_any_unimported = true

disallow_untyped_calls = true
disallow_untyped_defs = true

allow_untyped_globals = false
allow_redefinition = false

show_error_context = true
pretty = true
color_output = true
error_summary = true

html_report = "./reports/mypy/"

[tool.coverage.run]
branch = true
omit = ["src/ttnet/entrypoint.py"]

[tool.coverage.report]
fail_under = 0.00
precision = 2

[tool.pytest.ini_options]
addopts = "--cov-config pyproject.toml --cov src/ttnet --cov-report html:reports/pytest --cov-report xml:reports/coverage.xml --cov-report term-missing"
console_output_style = "progress"
required_plugins = "pytest-cov pytest-mock"
tmp_path_retention_policy = "none"
testpaths = ["tests"]
